Fix long webhooks content
