+++ src/main/java/me/alexpresso/zuninja/services/dispatch/DispatchService.java
@@ -1,9 +1,11 @@
 package me.alexpresso.zuninja.services.dispatch;
 
-import club.minnced.discord.webhook.send.WebhookEmbed;
+import club.minnced.discord.webhook.send.WebhookMessageBuilder;
+import me.alexpresso.zuninja.classes.projection.ProjectionSummary;
+import me.alexpresso.zuninja.exceptions.NodeNotFoundException;
 
 public interface DispatchService {
-    void dispatch(WebhookEmbed embed);
+    void dispatch(WebhookMessageBuilder webhookMessage);
 
-    void dispatch(WebhookEmbed embed, String discordId);
+    void dispatch(ProjectionSummary projectionSummary, String discordTag) throws NodeNotFoundException;
 }
